% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.int.tnorm.r
\name{conf.int.tnorm}
\alias{conf.int.tnorm}
\title{Confidence intervals for truncated normal}
\usage{
conf.int.tnorm(
  x0,
  level = 0.95,
  alpha = 1 - level,
  sd = 1,
  low = -Inf,
  upp = +Inf
)
}
\arguments{
\item{x0}{observed value}

\item{upp}{}
}
\value{
a vector of length 2
}
\description{
Confidence intervals for truncated normal
}
\details{
this function needs improvement.
}
\examples{
conf.int.tnorm(0)
conf.int.tnorm(0, low = -1, upp = +Inf)
}
