% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.fq.r
\name{contour.fq}
\alias{contour.fq}
\title{Contour plot for a quadratic form}
\usage{
\method{contour}{fq}(x, c = 0, u = c(0, 0), A, n = 201, branche = c(-1, 1))
}
\arguments{
\item{x}{value of the quadratic form}

\item{c}{constant term}

\item{u}{linear term}

\item{A}{symmetric 2x2 matrix}

\item{n}{number of points}

\item{branch}{which branch to trace (in the elliptic case)}
}
\value{
a list with components x, y
}
\description{
Contour plot for a quadratic form
}
\examples{
A <- matrix(c(1,0.5,0.5,1), 2, 2)
plot( contour.fq(1, A = A), type = "l", xlab = "x", ylab = "y")
for(v in 0.8**(1:40)) lines( contour.fq(v, A = A) )
}
