% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.lasso.r
\name{demo.lasso}
\alias{demo.lasso}
\title{A small demo of what the lasso does}
\usage{
demo.lasso(x, y, lambda)
}
\description{
A small demo of what the lasso does
}
\examples{
set.seed(3)
x <- cbind( c(rep(-1, 20), rep(1,20)), runif(40, -1, 1) )
x[,2] <- x[,2] + 0.4 * x[,1] 
y <- x \%*\% c(1,2) + rnorm(40)
y <- y - mean(y)
demo.lasso(x, y, lambda)
}
