% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.ridge.r
\name{demo.ridge}
\alias{demo.ridge}
\title{A small demo of what Ridge does}
\usage{
demo.ridge(x, y, lambda, xlim = c(-5, 5), ylim = c(-5, 5), add.grad = FALSE)
}
\description{
A small demo of what Ridge does
}
\examples{
set.seed(3)
x <- cbind( c(rep(-1, 20), rep(1,20)), runif(40, -1, 1) )
x[,2] <- x[,2] + 0.4 * x[,1] 
y <- x \%*\% c(1,2) + rnorm(40)
y <- y - mean(y)
demo.ridge(x, y, 25)
}
