% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.A.b.r
\name{get.A.b}
\alias{get.A.b}
\title{Conditions for model choice}
\usage{
get.A.b(X, M, S)
}
\arguments{
\item{X}{matrix of covariates}

\item{M}{indices of covariates kept in the model}

\item{S}{signs of the effects}
}
\description{
Conditions for model choice
}
\details{
Corresponds to prop 4.2. S can contain zeros for unpenalized variables.
the corresponding model is selected if \eqn{A y <= lambda B}.
}
