% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.interval.r
\name{get.interval}
\alias{get.interval}
\title{Get a truncation interval}
\usage{
get.interval(eta, y, A, B, lambda)
}
\arguments{
\item{eta}{direction to project y}

\item{y}{value of the vector y}

\item{A}{matrix}

\item{B}{vector}

\item{lambda}{regularization parameter}
}
\description{
Get a truncation interval
}
\details{
One should have \eqn{A y <= lambda B} for this to make sense.
Gives the bounds on eta' y, when y verifies this inequality. The bounds depend on y through z
which is the residual of the projection of y on eta. Cf lemma 5.1.
}
