% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.intervals.r
\name{get.intervals}
\alias{get.intervals}
\title{Get a truncation interval}
\usage{
get.intervals(eta, y, linear.constraints, lambda, sanitize = TRUE)
}
\arguments{
\item{eta}{direction to project y}

\item{y}{value of the vector y}

\item{linear.constraints}{As sent back by \code{get.linear.constraints}}

\item{lambda}{regularization parameter}
}
\description{
Get a truncation interval
}
\details{
Gives a list of bounds on eta' y, when y verifies this inequality. The bounds depend on y through z
which is the residual of the projection of y on eta. Cf lemma 5.1.
}
