% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.linear.constraints.r
\name{get.linear.constraints}
\alias{get.linear.constraints}
\title{Linear constraints for model choice}
\usage{
get.linear.constraints(X, M, S)
}
\arguments{
\item{X}{matrix of covariates}

\item{M}{indices of covariates kept in the model}

\item{S}{signs of the effects}
}
\description{
Linear constraints for model choice
}
\details{
Corresponds to prop 4.2. \code{S} can contain zeros for unpenalized variables.
the corresponding model is selected if \eqn{A y <= lambda B}.

If S is missing or contains \code{NA}s, then a list of linear constraints 
will be returned.
}
