% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.tnorm.r
\name{shiny.tnorm}
\alias{shiny.tnorm}
\title{Shiny demo for truncated normal confidence interval}
\usage{
shiny.tnorm(x0 = 0, low = -Inf, upp = Inf, sd = 1)
}
\description{
Shiny demo for truncated normal confidence interval
}
