% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm.r
\name{TNorm}
\alias{TNorm}
\alias{dtnorm}
\alias{ptnorm}
\alias{qtnorm}
\alias{rtnorm}
\title{Truncated normal}
\usage{
dtnorm(x, mean = 0, sd = 1, low = -Inf, upp = +Inf)

ptnorm(q, mean = 0, sd = 1, low = -Inf, upp = +Inf)

qtnorm(p, mean = 0, sd = 1, low = -Inf, upp = +Inf)

rtnorm(n, mean = 0, sd = 1, low = -Inf, upp = +Inf)
}
\description{
Truncated normal
}
